requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.8.6
python-dotenv==0.17.1
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
